更新unit test 更新swagger 將部分代碼改成符合swagger的結構

+++ src/main/java/com/momo/dylantest/mapper/postgres/CompanyStockPostgresMapper.java
@@ -45,5 +45,5 @@ public interface CompanyStockPostgresMapper {
     List<CompanyStockDto> findCompanyStockPage(@Param("offset") int offset, @Param("pageSize") int pageSize);
 
     @Select({"SELECT count(1) FROM company_stock "})
-    int findCompanyStockCount();
+    long findCompanyStockCount();
 }
